// libs/interfaces/enums.ts
export enum PostFeedTypes {
  POST = "POST",
  FEED = "FEED",
  ARTICLE = "ARTICLE",
  COMMENT = "COMMENT",
}

export enum AccreditationTypes {
  NONE = "NONE",
  PARTIAL = "PARTIAL",
  FULL = "FULL",
}

export enum ICriteria {
  GENERAL = "GENERAL",
  CITATIONS = "CITATIONS",
  INTERNATIONAL_OUTLOOK = "INTERNATIONAL_OUTLOOK",
  TEACHING = "TEACHING",
  RESEARCH = "RESEARCH",
  STUDENTS_EVALUATION = "STUDENTS_EVALUATION",
}

export enum StateTypes {
  ABIA = "ABIA",
  ANAMBRA = "ANAMBRA",
  ENUGU = "ENUGU",
  EBONYI = "EBONYI",
  IMO = "IMO",
}

export enum LecturerLevel {
  JUNIOR = "JUNIOR",
  SENIOR = "SENIOR",
}

export enum OwnerTypes {
  STATE = "STATE",
  FEDERAL = "FEDERAL",
  PRIVATE = "PRIVATE",
}

export enum AccountTypes {
  GUEST = "GUEST",
  ALUMNI = "ALUMNI",
  OWNER = "OWNER",
  ADMIN = "ADMIN",
  LECTURER = "LECTURER",
  GRADUATE = "GRADUATE",
  STUDENT = "STUDENT",
}

export enum AccountRoles {
  OWNER = "OWNER",
  ADMIN = "ADMIN",
  VC = "VC",
  HOD = "HOD",
  USER = "USER",
}

export enum Gender {
  NONE = "NONE",
  NOTSAY = "NOTSAY",
  MALE = "MALE",
  FEMALE = "FEMALE",
}

export enum MembershipTypes {
  LOCAL = "LOCAL",
  INTERNATIONAL = "INTERNATIONAL",
}

export enum StudentType {
  LOCAL = "LOCAL",
  INTERNATIONAL = "INTERNATIONAL",
}

export enum LecturerType {
  LOCAL = "LOCAL",
  INTERNATIONAL = "INTERNATIONAL",
}

export enum SchoolTypes {
  LOCAL = "LOCAL",
  INTERNATIONAL = "INTERNATIONAL",
}